Duplicate code from setup-envtest that can be easily reused
Since the error handling in the setup-envtest executable main path is
built on panics, recovers and os.Exit, it isn't ideal for re-using in a
version that downloads binaries within the test run.

However, much of the heavy lifting is performed by the remote, store and
versions packages, which _do_ properly return errors, and can readily be
reused in a new implementation.

This commit just duplicates the code from under tools/setup-envtest to
pkg/envtest, where we can build a new entry point that is easier to use
from within tests.